pyproject.toml
[tool.poetry]
name = "flask-wiki"
version = "0.2.4"
description = "Simple file-based wiki for Flask"
readme = "README.md"
authors = ["rero <info@rero.ch>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = ">=3.8.0"
bootstrap-flask = "*"
flask = "<3.0.0"
flask-wtf = "*"
babel = ">=2.9.1"
flask-babelex = "*"
wtforms = "<3.0.0"
werkzeug = ">=0.15"
jinja2 = ">=3.0.0"
markdown = "<3.4.0"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.5"
python-dotenv = "^1.0.0"
flask-bcrypt = "^1.0.1"
alembic = "^1.11.3"
flask-migrate = "^4.0.4"
email-validator = "^2.0.0.post2"

[tool.poetry.dev-dependencies]
check-manifest = "*"
ipython = "*"
autopep8 = "*"
poethepoet = "^0.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
extract_messages = "pybabel extract --project Flask-Wiki -F babel.cfg -k gettext -k lazy_gettext -o flask_wiki/translations/messages.pot flask_wiki"
init_catalog = "pybabel init -i flask_wiki/translations/messages.pot -d flask_wiki/translations -l "
compile_catalog = "pybabel compile -d flask_wiki/translations"
update_catalog = "pybabel update -i flask_wiki/translations/messages.pot -d flask_wiki/translations"
